backend/requirements/development.txt
# Development dependencies
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-watch==4.2.0
factory-boy==3.3.0
faker==22.0.0
locust==2.32.0

# Code quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
django-stubs==4.2.6

# Debugging
django-debug-toolbar==4.2.0
ipython==8.18.1
django-extensions==3.2.3

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development server
watchfiles==0.21.0

# For OCR/NER features in development, install additional dependencies:
# pip install -r ml.txt
# Then download spaCy German model: python -m spacy download de_core_news_lg
